requirements.txt
# Core dependencies
crewai[tools]>=0.86.0
sqlalchemy>=2.0.0
pydantic>=2.0.0

# LLM providers
groq>=0.4.0

# UI
streamlit>=1.30.0
rich>=13.0.0

# Utilities
python-dotenv>=1.0.0
tabulate>=0.9.0

# Note: sqlite3 is part of Python standard library (no pip install needed)
google-generativeai>=0.3.0
litellm>=1.0.0

# API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0

# PostgreSQL support (for Supabase/cloud deployments)
psycopg2-binary>=2.9.0
